+++ team_management_app/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:team_management_app/allscreen.dart';
 import 'package:team_management_app/screen/entireappbar.dart';
-import 'screen/githublogin.dart'; // GitHub 로그인 위젯을 포함하기 위한 임포트
+import 'screen/loginandsignup/githublogin.dart'; // GitHub 로그인 위젯을 포함하기 위한 임포트
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 void main() async {
